Moved classes to base modules
