+++ libdirections-offboard/src/main/java/com/mapbox/navigation/route/offboard/MapboxOffboardRouter.kt
@@ -48,6 +48,10 @@ class MapboxOffboardRouter(
         })
     }
 
+    override fun getRoute(origin: Point, waypoints: List<Point>, listener: Router.RouteListener) {
+
+    }
+
     override fun cancel() {
         navigationRoute?.cancelCall()
         navigationRoute = null
