+++ libnavigation-base/src/main/java/com/mapbox/navigation/base/route/Router.kt
@@ -18,3 +18,4 @@ interface Router {
         fun onFailure(throwable: Throwable)
     }
 }
+
